requirements.txt
numpy~=1.24.4
rasterio~=1.3.9
torch~=2.0.1
Pillow~=8.4.0
joblib~=1.3.2
torchvision~=0.15.2
pandas~=2.0.3
matplotlib~=3.7.5
scikit-learn~=1.3.2
scikit-image~=0.19.3
kornia~=0.6.12
torchgeo
tabulate~=0.9.0